[FPGA] Fix REFCLK issue for iBERT GTM property
@@ fix iBERT GTM property: REFCLK issue in bxu_ibert_setGTM_config.tcl @@
- Now support NRZ: 25G / 26G, PAM4: 51G / 53G / 56G / 106G / 112G

- The Versal Transceiver Bridge IP, has configurations for Ethernet / ILKN (Interlaken), NRZ / PAM4, 25G / ... / 112G.
  However, these PRESET configuration won't work by simply Vivado GUI's selection
  To get it work, we should reference the XilinxCEDStore's iBERT example code.
  - TX/RX_REFCLK_FREQUENCY:  Vivado GUI PRESET = 161.1328125 (failed), CEDStore-iBERT-example = 156.25 (OK)
  - TX/RXPROGDIV_FREQ_VAL, TX/RX_USER_DATA_WIDTH: let Vivado GUI update for us
  - INS_LOSS_NYQ (Insertion loss at Nyquist): can be modified to fine-tuning signal quality. Ex. 35dB

@@ ChipScoPy/HPC_Cable_BERT_YK_Scan.py changes @@
- QTableWidget Column: "Samples" -> "YK-Samples", "Link-Samples"
- QTableWidget Column: "comments" supports Pandas statistical data
  For "No Link": shows more details of link status
  For normal link: shows BER/SNR statistical data

- supports QtWidgets::setWindowTitle() with customized font size & color

@@ PDI file paths adjustment for all supported rates: 25G / ... / 112G @@
- ChipScoPy python script & links update for all supported rates

@@ misc @@
- Unified the naming for Link Channel Mapping file, with theseconnection scheme:
  XConnected (XConn) / SelfLooped (SLoop)

@@ Minor modifications on BASH script for FPGA @@
- Invokes of Xilinx hw_server / cs_server on bin/FPGA_Server_Xilinx.sh only.
  While bin/chipScopy.sh only need to source FPGA_Server_Xilinx.sh

+++ server_PC/bin/FPGA_Server_Xilinx.sh
@@ -1,18 +1,74 @@
 #!/bin/bash
 #----------------------------------------------------------------------------
+# export HWSERV_RESTART=True           # env-var to indicate whether to kill previous hw_server or not
+# export LOCAL_FPGA_HW=True            # env-var for remote ChipScoPy/Jupyter execution
+#----------------------------------------------------------------------------
 JTAG_PORT=${JTAG_PORT:=3121}           # externally defined export JTAG_PORT, default 3121
 JTAG_PORT=${1:-$JTAG_PORT}             # input parameter %1 for JTAG_PORT
 
-# kill all previous instances
-sudo killall -9 cs_server
-sudo killall -9 hw_server
-rm -f /tmp/FPGA-$JTAG_PORT.*   # remove old LOGFILE
+rm -f /tmp/FPGA-*.log                  # remove old LOGFILE
 
-TOOLPATH=/opt/Xilinx/tools/Vivado_Lab/2023.2/bin
-[ ! -d "$TOOLPATH" ] && TOOLPATH=/opt/Xilinx/tools/Vivado/2023.2/bin
-[ ! -d "$TOOLPATH" ] && TOOLPATH=/fpga_share/Xilinx_tools/Vivado_Lab/2023.2/bin
-[ ! -d "$TOOLPATH" ] && TOOLPATH=/fpga_share/Xilinx_tools/Vivado/2023.2/bin
-cd $TOOLPATH
+if [[ `uname -r` = *"microsoft-standard-WSL2"* ]]; then
+	#export CS_SERVER_URL="TCP:10.20.2.146:3042"
+	#export HW_SERVER_URL="TCP:10.20.2.146:3121"
 
-./cs_server & disown
-./hw_server -d -levents,eventcore,protocol,discovery,tcflog,jtag2,jtag,xvc,pcie -L/tmp/FPGA-$JTAG_PORT.`date +%F_%H%M`.log -stcp::$JTAG_PORT
+	# /mnt/c/Windows/System32/cmd.exe /c 'tasklist ' | grep hw_server
+
+	if [ "$LOCAL_FPGA_HW" = "True" ]; then
+		#----------------------------------------------------------------------------
+		if ! /mnt/c/Windows/System32/cmd.exe /c 'tasklist ' | grep hw_server; then
+			echo -e "\nRestart hw_server"
+			/mnt/c/Windows/System32/cmd.exe /c 'C:\Xilinx\Vivado\2023.2\bin\hw_server.bat' & disown
+		fi
+
+		#----------------------------------------------------------------------------
+		echo -e "\nRestart cs_server"
+		/mnt/c/Windows/System32/cmd.exe /c 'taskkill /IM cs_server.exe /F'
+		/mnt/c/Windows/System32/cmd.exe /c 'C:\Xilinx\Vivado\2023.2\bin\cs_server.bat' & disown
+	fi
+else
+	TOOLPATH=""
+	[ ! -d "$TOOLPATH" ] && TOOLPATH=/opt/Xilinx/tools/Vivado/2023.2/bin
+	[ ! -d "$TOOLPATH" ] && TOOLPATH=/opt/Xilinx/tools/Vivado_Lab/2023.2/bin
+	[ ! -d "$TOOLPATH" ] && TOOLPATH=/fpga_share/Xilinx_tools/Vivado/2023.2/bin
+	[ ! -d "$TOOLPATH" ] && TOOLPATH=/fpga_share/Xilinx_tools/Vivado_Lab/2023.2/bin
+
+	if [ "$LOCAL_FPGA_HW" = "True" ] && [ -d "$TOOLPATH" ]; then
+		cd $TOOLPATH
+
+		#----------------------------------------------------------------------------
+		[ "$HWSERV_RESTART" = "True" ]  && sudo killall -9 hw_server     # kill previous instances
+		if ! ps -ef | egrep '[h]w_server' > /dev/null; then
+			echo -e "\nRestart hw_server"
+			./hw_server  -d -L/tmp/FPGA-hw_$JTAG_PORT.`date +%F_%H%M`.log -stcp::$JTAG_PORT -levents,eventcore,protocol,discovery,tcflog,jtag2,jtag,xvc,pcie
+			#./hw_server -d -L/tmp/FPGA-$JTAG_PORT.`date +%F_%H%M`.log -stcp::$JTAG_PORT
+			#./hw_server -d -L/tmp/FPGA-$JTAG_PORT.`date +%F_%H%M`.log -stcp::$JTAG_PORT -e "set xvc-servers $XVC_SERVER:localhost:3000"
+		fi
+
+		#----------------------------------------------------------------------------
+		# Usage: cs_server [OPTIONS]
+		# 
+		# Options:
+		#   -L, --log-file <file>           enable logging, use -L- to send log to
+		#                                   stdout
+		#   -l, --log-domains <domains>     enable logging based on domain. Format -
+		#                                   "<domains separated by comma>:<level>". The
+		#                                   level should be one of the following: TRACE/
+		#                                   DEBUG/INFO/SUCCESS/WARNING/ERROR/CRITICAL
+		#   -s <url>                        set agent listening port and protocol
+		#                                   [default: TCP::3042]
+		#   -d, --daemon-mode               run in daemon mode
+		#   -I, --idle-timeout <idle-seconds>
+		#                                   exit if there are no connections for the
+		#                                   specified time
+		#   -i, --interactive               run in interactive mode
+		#----------------------------------------------------------------------------
+		sudo killall -9 cs_server     # kill previous instances
+		if ! ps -ef | egrep '[c]s_server' > /dev/null; then
+			echo -e "\nRestart cs_server"
+			./cs_server -L /tmp/FPGA-cs_server.`date +%F_%H%M`.log -d  & disown
+			# ./cs_server & disown
+		fi
+	fi
+fi
+
